Bring test_eventer server into project, register pusher name
Config files for dev and test have been added
to allow using an external and test eventer endpoint.

The pusher file registers a name: this change was introduced to allow
supervisors to be added and to allow dets tables to be used.

A fake eventer has been added to run when tests are run.